Add bin feature to lib target
